vfs: fix incorrect modtime when uploading files with --vfs-cache-mode full
Before this change when using --vfs-cache-mode full the modtime would
only update after the object had been written to storage.

This caused nano to think that the object had changed and produce an
error when saving a file.

This change reads the ModTime from the dirty cache file if one is
available in preference to the old modtime from the previous object.

Fixes #4763